remove fast objects
